Move init_cmd within own package
Fixes #129
