+++ cmd/check_available_memory/cmd/root.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/ncr-devops-platform/nagiosfoundation/cmd/initcmd"
 	"github.com/ncr-devops-platform/nagiosfoundation/lib/app/nagiosfoundation"
 	"github.com/spf13/cobra"
 )
@@ -28,7 +29,7 @@ issue a WARNING response. Otherwise, an OK response is issued.`,
 		},
 	}
 
-	nagiosfoundation.AddVersionCommand(rootCmd)
+	initcmd.AddVersionCommand(rootCmd)
 
 	rootCmd.Flags().IntVarP(&warning, "warning", "w", 85, "the memory threshold to issue a warning alert")
 	rootCmd.Flags().IntVarP(&critical, "critical", "c", 95, "the memory threshold to issue a critical alert")
